requirements.txt
flask 
pymupdf 
sentence-transformers 
faiss-cpu 
transformers                                                              